refactor(tela-inicial): muda nomes dos componentes e das paginas em geral
