+++ src/components/main-page/Text-box.tsx
